fixed double->float8
git-svn-id: http://svn.code.sf.net/p/rts-2/code/trunk/rts-2@6284 65ab8f4e-f147-0410-b3a9-f14133ecfe55

+++ src/sql/auger/rel_0_8_0.sql
@@ -5,8 +5,8 @@ DROP TABLE auger;
 CREATE TABLE auger (
 	auger_t3id	integer PRIMARY KEY NOT NULL,
 	auger_date	timestamp NOT NULL,
-	ra		double NOT NULL,
-	dec		double NOT NULL,
+	ra		float8 NOT NULL,
+	dec		float8 NOT NULL,
 	-- now all fields from message..
 
 	eye		integer NOT NULL, -- FD eye Id
